+++ src/main/java/com/kard/api/resources/organizations/placements/AsyncRawPlacementsClient.java
@@ -178,6 +178,14 @@ public class AsyncRawPlacementsClient {
                 .addPathSegments("v2/issuers")
                 .addPathSegment(organizationId)
                 .addPathSegments("placements");
+        if (request.getFilterType().isPresent()) {
+            QueryStringMapper.addQueryParameter(
+                    httpUrl, "filter[type]", request.getFilterType().get(), false);
+        }
+        if (request.getFilterName().isPresent()) {
+            QueryStringMapper.addQueryParameter(
+                    httpUrl, "filter[name]", request.getFilterName().get(), false);
+        }
         if (request.getPageAfter().isPresent()) {
             QueryStringMapper.addQueryParameter(
                     httpUrl, "page[after]", request.getPageAfter().get(), false);
@@ -215,6 +223,11 @@ public class AsyncRawPlacementsClient {
                     }
                     try {
                         switch (response.code()) {
+                            case 400:
+                                future.completeExceptionally(new InvalidRequest(
+                                        ObjectMappers.JSON_MAPPER.readValue(responseBodyString, ErrorResponse.class),
+                                        response));
+                                return;
                             case 401:
                                 future.completeExceptionally(new UnauthorizedError(
                                         ObjectMappers.JSON_MAPPER.readValue(responseBodyString, ErrorResponse.class),